Refactor workflow notification code
The init function contains only dispatches requests to propper files.
All functions for handling data and notifications were moved to
data_handler and notification_handler files.
